LayoutTests/imported/w3c/web-platform-tests/css/css-values/if-function-revert-rule-expected.txt


FAIL The revert-rule works as a branch in if() assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
FAIL The revert-rule works as a branch in if(), earlier return assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
